Add a basic CLI for Basilisp REPL
